% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.CibersortxTpm.R
\name{Get.CibersortxTpm}
\alias{Get.CibersortxTpm}
\title{Title}
\usage{
Get.CibersortxTpm(
  seuart_object = seuart_object,
  celltype_varname = celltype_varname,
  select_allcells = F,
  gsub_cellname = gsub_cellname,
  specified_cells = specified_cells,
  cl = 1,
  write_filename = NULL,
  gsub_string = gsub_string,
  sep = "\\t",
  row.names = F,
  col.names = T
)
}
\arguments{
\item{col.names}{}
}
\description{
Title
}
